requirements.txt
google-generativeai
python-dotenv
chromadb
colorama
pydantic
